// src/app/features/admin-panel/components/inquiry-form/@enum/form-enum.ts
export enum InquiryFormName {
  INQUIRY_NAME = 'inquiryName',
  QUESTIONS = 'questions'
}

export enum ShortTextQuestionFormName {
  QUESTION = 'question',
  ANSWER = 'answer',
  ID = 'id'
}

export enum SingleSelectQuestionFormName {
  ID = 'id',
  ANSWERS = 'answers',
  QUESTION = 'question'
}

export enum MultiSelectQuestionFormName {
  QUESTION = 'question',
  ANSWERS = 'answers',
  ID = 'id'
}

export enum TypeQuestion {
  TYPE = 'type'
}

export enum ScaleSelectQuestionFormName {
  QUESTION = 'question',
  ID = 'id',
  MIN_VALUE = 'min-value',
  MAX_VALUE = 'max-value',
  STEP_SIZE = 'step-size'
}